Ability to override Nethermind flags
I need my Nethermind node to run the `Trace` module, as well as include
additional values into `JsonRpc.EnabledModules`. Adding extra duplicate
flags through `EC_ADDITIONAL_FLAGS`. This is the next best thing I could
think of.

With this change, users should be able to add the whole
`JsonRpc.EnabledModules` and `JsonRpc.AdditionalUrls` flags to
`EC_ADDITIONAL_FLAGS`, and the script will fill in the default values if
they're not present.
In this approach, it's up to make sure they still include the correct modules
and urls that RocketPool requires by default