Properly handle case of unknown external IP
The `smartnode` applies `shellescape.Quote` to environment variables.
When no external IP address is known (e.g., in an intermittent outage),
`EXTERNAL_IP` is set to `shellescape.Quote("")` which is `''`.
However, in `smartnode-install` we check for an actually empty string.
Our check incorrectly assumes that `''` is a valid IP address, and then
proceeds to try and start Nimbus using `--nat=extip:''`, which fails.
This is fixed by also checking that `EXTERNAL_IP` is not set to `''`.